scrolling: initial reverse scrolling support - no scroll regions

+++ grid.h
@@ -3,6 +3,6 @@
 #include <stddef.h>
 #include "terminal.h"
 
-struct cell *grid_get_range(struct grid *grid, size_t start, size_t *length);
-void grid_memset(struct grid *grid, size_t start, int c, size_t length);
-void grid_memmove(struct grid *grid, size_t dst, size_t src, size_t length);
+struct cell *grid_get_range(struct grid *grid, int start, int *length);
+void grid_memset(struct grid *grid, int start, int c, int length);
+void grid_memmove(struct grid *grid, int dst, int src, int length);
